# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "RainForge API",
        "description": "Complete Rooftop Rainwater Harvesting Platform API\n\n## Features\n- **Assessment**: Instant RWH assessment with 3 scenarios\n- **PDF Generation**: Engineer-signed reports with QR verification\n- **Auction**: Reverse auction for installer bidding\n- **Allocation**: Smart installer allocation (RPI-weighted)\n- **Escrow**: Milestone-based payment release\n- **Verification**: Fraud-resistant geo-verified photo uploads\n- **Monitoring**: Real-time IoT telemetry\n- **Public Dashboard**: Ward-level transparency reporting",
        "version": "1.0.0",
        "contact": {
            "name": "RainForge Team",
            "email": "support@rainforge.in"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development"
        }
    ],
...
    "paths": {
...
        "/escrow/{job_id}/release": {
            "post": {
                "tags": [
                    "Escrow"
                ],
                "summary": "Release Escrow Milestone",
                "description": "Release payment for a completed milestone (10% design, 70% install, 20% verify).",
                "parameters": [
                    {
                        "name": "job_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EscrowReleaseRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Milestone released"
                    },
                    "400": {
                        "description": "Milestone already released"
                    },
                    "404": {
                        "description": "Job or escrow not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EscrowReleaseRequest": {
                "type": "object",
                "required": [
                    "milestone_id"
                ],
                "properties": {
                    "milestone_id": {
                        "type": "string"
                    },
                    "verification_id": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}